+++ src/main/java/com/example/demo/controller/HeroesController.java
@@ -15,4 +15,6 @@ public class HeroesController {
     public String getHeroes(@PathVariable int id){
         return  heroesService.selHero(id).toString();
     }
-}
+
+    }
+
