+++ src/main/resources/mapping/HeroesMapper.xml
@@ -1,15 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.example.demo.mapper.HeroesMapper">
-    <resultMap id="BaseResultMap" type="com.example.demo.entity.Heroes">
+    <resultMap id="BaseResultMap" type="com.example.demo.entity.Hero">
         <result column="id" jdbcType="INTEGER" property="id" />
         <result column="title" jdbcType="VARCHAR" property="heroTitle" />
         <result column="name" jdbcType="VARCHAR" property="heroName" />
         <result column="types" jdbcType="VARCHAR" property="heroTypes" />
     </resultMap>
 
-    <select id="selHero" resultType="com.example.demo.entity.Heroes">
+    <select id="selHero" resultType="com.example.demo.entity.Hero">
         select * from heroes where id = #{id}
+    </select>
+    <select id="selHeroRandom" resultType="com.example.demo.entity.Hero">
+        select * from heroes where id =#{id}
     </select>
 
 </mapper>
